moved commands.h file to network folder
